+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+*   Add write support to Spanner.
+
 ## 1.1.1 - 2025-12-18
 
 *   Prefix column names with table name to avoid ambiguity